pyproject.toml
[project]
name = "stock-rag"
version = "0.1.0"
description = "10-K RAG Pipeline POC for financial document querying"
authors = [
    {name = "Jacques Camier", email = "jacques@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "psycopg2-binary>=2.9",
    "pgvector>=0.2.0",
    "openai>=1.0",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "pydantic>=2.0",
    "python-decouple>=3.8",
    "langfuse>=3.0",
    "tiktoken>=0.5",
    "numpy>=1.24",
    "pandas>=2.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.24",
    "gunicorn>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "black>=23.0",
    "mypy>=1.5",
    "pre-commit>=3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=rag_pipeline --cov-report=html --cov-report=term-missing"
